moved windoweddataextractor to segmenting
