+++ src/paradigma/segmenting.py
@@ -4,7 +4,6 @@ import numpy as np
 from typing import List
 from paradigma.constants import DataColumns
 
-
 import numpy as np
 
 def tabulate_windows(
@@ -310,3 +309,88 @@ def categorize_segments(df, fs, format='timestamps', window_step_length_s=None):
 
     # Apply categorization to the DataFrame
     return df[DataColumns.SEGMENT_NR].map(segment_sizes).map(categorize).astype('category')
+
+class WindowedDataExtractor:
+    """
+    A utility class for extracting specific column indices and slices 
+    from a list of windowed column names.
+
+    Attributes
+    ----------
+    column_indices : dict
+        A dictionary mapping column names to their indices.
+
+    Methods
+    -------
+    get_index(col)
+        Returns the index of a specific column.
+    get_slice(cols)
+        Returns a slice object for a range of consecutive columns.
+    """
+
+    def __init__(self, windowed_cols):
+        """
+        Initialize the WindowedDataExtractor.
+
+        Parameters
+        ----------
+        windowed_cols : list of str
+            A list of column names in the windowed data.
+
+        Raises
+         ------
+        ValueError
+            If the list of `windowed_cols` is empty.
+        """
+        if not windowed_cols:
+            raise ValueError("The list of windowed columns cannot be empty.")
+        self.column_indices = {col: idx for idx, col in enumerate(windowed_cols)}
+
+    def get_index(self, col):
+        """
+        Get the index of a specific column.
+
+        Parameters
+        ----------
+        col : str
+            The name of the column to retrieve the index for.
+
+        Returns
+        -------
+        int
+            The index of the specified column.
+
+        Raises
+        ------
+        ValueError
+            If the column is not found in the `windowed_cols` list.
+        """
+        if col not in self.column_indices:
+            raise ValueError(f"Column '{col}' not found in windowed_cols.")
+        return self.column_indices[col]
+
+    def get_slice(self, cols):
+        """
+        Get a slice object for a range of consecutive columns.
+
+        Parameters
+        ----------
+        cols : list of str
+            A list of consecutive column names to define the slice.
+
+        Returns
+        -------
+        slice
+            A slice object spanning the indices of the given columns.
+
+        Raises
+        ------
+        ValueError
+            If one or more columns in `cols` are not found in the `windowed_cols` list.
+        """
+        if not all(col in self.column_indices for col in cols):
+            missing = [col for col in cols if col not in self.column_indices]
+            raise ValueError(f"The following columns are missing from windowed_cols: {missing}")
+        start_idx = self.column_indices[cols[0]]
+        end_idx = self.column_indices[cols[-1]] + 1
+        return slice(start_idx, end_idx)
